@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,16 +27,16 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  firebase_core: ^0.5.3
-  firebase_auth: ^0.18.4+1
-  cloud_firestore: ^0.14.4
-  google_fonts: ^1.1.1
-  font_awesome_flutter: ^8.11.0
-  intl: ^0.16.1
-  qr_flutter: ^3.2.0
-  flutter_barcode_scanner: ^1.0.1
-  modal_progress_hud: ^0.1.3
+  cupertino_icons: ^1.0.5
+  firebase_core: ^2.2.0
+  firebase_auth: ^4.1.2
+  cloud_firestore: ^4.0.5
+  google_fonts: ^3.0.1
+  font_awesome_flutter: ^10.2.1
+  intl: ^0.17.0
+  qr_flutter: ^4.0.0
+  flutter_barcode_scanner: ^2.0.0
+  modal_progress_hud_nsn: ^0.3.0
 
 dev_dependencies:
   flutter_test:
